fix: more cleanup of code
