+++ custom_components/hsem/custom_sensors/HSEMIntegrationSensor.py
@@ -1,0 +1,22 @@
+from homeassistant.components.integration.sensor import IntegrationSensor
+from custom_components.hsem.entity import HSEMEntity
+
+class HSEMIntegrationSensor(IntegrationSensor, HSEMEntity):
+    """Custom Integration Sensor with device_info."""
+
+    def __init__(self, *args, id=None, config_entry=None, **kwargs):
+        IntegrationSensor.__init__(self, *args, **kwargs)
+        HSEMEntity.__init__(self, config_entry)
+        self._unique_id = id
+
+    @property
+    def unique_id(self):
+        return self._unique_id
+
+    async def async_added_to_hass(self):
+        """Handle the sensor being added to Home Assistant."""
+        await super().async_added_to_hass()
+
+    async def async_will_remove_from_hass(self):
+        """Entity being removed from hass."""
+        await super().async_will_remove_from_hass()
