+++ custom_components/hsem/sensor.py
@@ -1,15 +1,6 @@
-import logging
-
-from custom_components.hsem.const import DOMAIN
-from custom_components.hsem.custom_sensors.house_consumption_power_sensor import (
-    HouseConsumptionPowerSensor,
-)
-from custom_components.hsem.custom_sensors.working_mode_sensor import WorkingModeSensor
+from custom_components.hsem.custom_sensors.HSEMHouseConsumptionPowerSensor import HSEMHouseConsumptionPowerSensor
+from custom_components.hsem.custom_sensors.HSEMWorkingModeSensor import HSEMWorkingModeSensor
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from custom_components.hsem.utils.misc import get_config_value
-
-_LOGGER = logging.getLogger(__name__)
-
 
 async def async_setup_entry(hass, config_entry, async_add_entities: AddEntitiesCallback) -> None:
     """
@@ -30,7 +21,7 @@ async def async_setup_entry(hass, config_entry, async_add_entities: AddEntitiesC
     """
 
     # Setup working mode sensor
-    working_mode_sensor = WorkingModeSensor(config_entry)
+    working_mode_sensor = HSEMWorkingModeSensor(config_entry)
 
     async_add_entities([working_mode_sensor])
 
@@ -39,32 +30,8 @@ async def async_setup_entry(hass, config_entry, async_add_entities: AddEntitiesC
     for hour in range(24):
         hour_start = hour
         hour_end = (hour + 1) % 24
-        sensor = HouseConsumptionPowerSensor(config_entry, hour_start, hour_end, async_add_entities)
+        sensor = HSEMHouseConsumptionPowerSensor(config_entry, hour_start, hour_end, async_add_entities)
         power_sensors.append(sensor)
 
     # Add sensors to Home Assistant
     async_add_entities(power_sensors)
-
-    # Store reference to the platform to handle unloads later
-    #if DOMAIN not in hass.data:
-    #    hass.data[DOMAIN] = {}
-    #hass.data[DOMAIN][config_entry.entry_id] = async_add_entities
-
-#async def async_unload_entry(hass, entry):
-    """
-    Handle unloading of an entry.
-
-    This function is responsible for unloading a specific entry from the Home Assistant instance.
-    It retrieves the platform associated with the entry from the hass data and attempts to remove it.
-
-    Args:
-        hass (HomeAssistant): The Home Assistant instance.
-        entry (ConfigEntry): The configuration entry to unload.
-
-    Returns:
-        bool: True if the entry was successfully removed, False otherwise.
-    """
-#    platform = hass.data[DOMAIN].get(entry.entry_id)
-#    if platform:
-#        return await platform.async_remove_entry(entry)
-#    return False
